backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
drf-yasg==1.21.7
redis==5.0.1
django-cors-headers==4.3.1